requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.5.24
dataclasses-json==0.5.7
decorator==5.1.1
duckdb==0.8.0
entrypoints==0.4
faiss-cpu==1.7.4
fastapi==0.95.2
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.173
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
posthog==3.0.1
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.8.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.14
starlette==0.27.0
streamlit==1.22.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.29.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.3
wikipedia==1.4.0
yarl==1.9.2
youtube-transcript-api==0.6.0
zipp==3.15.0
zstandard==0.21.0
